build: add function to play 5 rounds
Added:
  - playGame() function encapsulating previously created score variables
    and playRound() function.
  - for loop to recall playRound() five times.
  - winner declaration at the end of all five rounds based on players'
    final score.

Moved previously global-scoped compChoice and humChoice variables to
their respective functions, leading to the reorganizing of said
functions by adding separate console.log() and return statements so as
to not return entire strings.
The return statements now fit the playRound() conditions of a single
word, and the strings have been moved to a console.log() statement.

Created new compSelection and humSelection variables recalling the main
getComputerChoice() and getHumanChoice() functions.